Share CLI exit status type
