+++ crates/karva/src/lib.rs
@@ -1,9 +1,10 @@
 use std::ffi::OsString;
-use std::process::{ExitCode, Termination};
 
 use anyhow::Context;
 use clap::Parser;
 use karva_cli::{Args, Command};
+
+pub use karva_cli::ExitStatus;
 
 mod commands;
 mod utils;
@@ -42,29 +43,5 @@ fn run(f: impl FnOnce(Vec<OsString>) -> Vec<OsString>) -> anyhow::Result<ExitSta
             commands::show_config::show_config(show_config_args)
         }
         Command::Version => commands::version::version().map(|()| ExitStatus::Success),
-    }
-}
-
-#[derive(Copy, Clone)]
-pub enum ExitStatus {
-    /// Checking was successful and there were no errors.
-    Success = 0,
-
-    /// Checking was successful but there were errors.
-    Failure = 1,
-
-    /// Checking failed.
-    Error = 2,
-}
-
-impl Termination for ExitStatus {
-    fn report(self) -> ExitCode {
-        ExitCode::from(self as u8)
-    }
-}
-
-impl ExitStatus {
-    pub fn to_i32(self) -> i32 {
-        self as i32
     }
 }
