+++ crates/karva_worker/src/bin/main.rs
@@ -1,4 +1,5 @@
-use karva_worker::cli::{ExitStatus, karva_worker_main};
+use karva_cli::ExitStatus;
+use karva_worker::cli::karva_worker_main;
 
 fn main() -> ExitStatus {
     karva_worker_main(|args| args)
