+++ crates/karva_worker/src/cli.rs
@@ -1,12 +1,11 @@
 use std::ffi::OsString;
-use std::process::{ExitCode, Termination};
 
 use anyhow::Context as _;
 use camino::Utf8PathBuf;
 use clap::Parser;
 use fs_err as fs;
 use karva_cache::{RunCache, RunHash};
-use karva_cli::{SubTestCommand, Verbosity};
+use karva_cli::{ExitStatus, SubTestCommand, Verbosity};
 use karva_diagnostic::{DummyReporter, Reporter, TestCaseReporter};
 use karva_logging::{Printer, StatusLevel, set_colored_override, setup_tracing};
 use karva_metadata::RunIgnoredMode;
@@ -47,29 +46,6 @@ impl Args {
     }
 }
 
-#[derive(Copy, Clone)]
-pub enum ExitStatus {
-    /// Checking was successful and there were no errors.
-    Success = 0,
-
-    /// Checking was successful but there were errors.
-    Failure = 1,
-
-    /// Checking failed.
-    Error = 2,
-}
-
-impl Termination for ExitStatus {
-    fn report(self) -> ExitCode {
-        ExitCode::from(self as u8)
-    }
-}
-
-impl ExitStatus {
-    pub fn to_i32(self) -> i32 {
-        self as i32
-    }
-}
 pub fn karva_worker_main(f: impl FnOnce(Vec<OsString>) -> Vec<OsString>) -> ExitStatus {
     run(f).unwrap_or_else(|error| {
         if karva_logging::error_chain_contains_broken_pipe(error.chain()) {
